don't skip ShiftChanges when timing is tight
For a pair of closely-spaced ShiftChanges where the first was calculated
from one day, and the second from another, the second was often skipped
because after generating a new day of shiftchanges, the next relevant
shiftchange was sought using the .trueAt time of the last triggered
shiftchange, which could be considerably later than its nominal .At time
This was especially so for dilated clocks, but would also happen with
back-to-back shifts.

Now, use the ordering of the .At fields to find the next ShiftChange so
that we don't skip any.